LRCI-3783 Move getTestResults logic for downstream builds into the BaseParentBuild
